test(openai-compat): AsyncOpenAI 构造替身收编到共享记录器
8 个测试文件散落的 patch("lib.openai_shared.AsyncOpenAI") 统一走 tests/fakes.py
的 captured_openai_clients，agnes 的本地同形 helper 折入共享版。
