+++ tests/unit/lib/text_backends/test_agnes_text_backend.py
@@ -8,9 +8,6 @@ AgnesTextBackend 复用 OpenAITextBackend 的原生 + Instructor 降级流水线
 from __future__ import annotations
 
 import json
-from collections.abc import Iterator
-from contextlib import contextmanager
-from typing import Any
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
@@ -18,19 +15,7 @@ from pydantic import BaseModel
 
 from lib.providers import PROVIDER_AGNES
 from lib.text_backends.base import TextCapability, TextGenerationRequest
-
-
-@contextmanager
-def _recorded_openai_clients() -> Iterator[list[dict[str, Any]]]:
-    """AsyncOpenAI 构造的记录器：收下建客户端的参数，回一个空替身。"""
-    created: list[dict[str, Any]] = []
-
-    def _create(**kwargs: Any) -> Any:
-        created.append(kwargs)
-        return AsyncMock()
-
-    with patch("lib.openai_shared.AsyncOpenAI", _create):
-        yield created
+from tests.fakes import captured_openai_clients
 
 
 def _make_mock_response(content="Hello", input_tokens=10, output_tokens=5):
@@ -59,7 +44,7 @@ class _PersonSchema(BaseModel):
 
 class TestConstruction:
     def test_name_and_default_model(self):
-        with patch("lib.openai_shared.AsyncOpenAI"):
+        with captured_openai_clients():
             from lib.text_backends.agnes import AgnesTextBackend
 
             backend = AgnesTextBackend(api_key="sk")
@@ -67,14 +52,14 @@ class TestConstruction:
             assert backend.model == "agnes-2.0-flash"
 
     def test_custom_model(self):
-        with patch("lib.openai_shared.AsyncOpenAI"):
+        with captured_openai_clients():
             from lib.text_backends.agnes import AgnesTextBackend
 
             backend = AgnesTextBackend(api_key="sk", model="agnes-2.0-pro")
             assert backend.model == "agnes-2.0-pro"
 
     def test_capabilities_text_and_structured_no_vision(self):
-        with patch("lib.openai_shared.AsyncOpenAI"):
+        with captured_openai_clients():
             from lib.text_backends.agnes import AgnesTextBackend
 
             backend = AgnesTextBackend(api_key="sk")
@@ -84,7 +69,7 @@ class TestConstruction:
             assert TextCapability.VISION not in backend.capabilities
 
     def test_missing_api_key_raises(self):
-        with patch("lib.openai_shared.AsyncOpenAI"):
+        with captured_openai_clients():
             from lib.text_backends.agnes import AgnesTextBackend
 
             with pytest.raises(ValueError, match="Agnes API Key"):
@@ -92,14 +77,14 @@ class TestConstruction:
 
     def test_base_url_normalized_to_v1(self):
         # 用户填 host（无 /v1 后缀）→ 归一化为 {host}/v1 后传给 SDK。
-        with _recorded_openai_clients() as created:
+        with captured_openai_clients() as created:
             from lib.text_backends.agnes import AgnesTextBackend
 
             AgnesTextBackend(api_key="sk", base_url="https://apihub.agnes-ai.com")
         assert [c["base_url"] for c in created] == ["https://apihub.agnes-ai.com/v1"]
 
     def test_default_base_url_when_unset(self):
-        with _recorded_openai_clients() as created:
+        with captured_openai_clients() as created:
             from lib.text_backends.agnes import AgnesTextBackend
 
             AgnesTextBackend(api_key="sk")
@@ -111,7 +96,7 @@ class TestGenerate:
         mock_client = AsyncMock()
         mock_client.chat.completions.create = AsyncMock(return_value=_make_mock_response("Test output", 15, 8))
 
-        with patch("lib.openai_shared.AsyncOpenAI", return_value=mock_client):
+        with captured_openai_clients(mock_client):
             from lib.text_backends.agnes import AgnesTextBackend
 
             backend = AgnesTextBackend(api_key="sk")
@@ -127,7 +112,7 @@ class TestGenerate:
         mock_client = AsyncMock()
         mock_client.chat.completions.create = AsyncMock(return_value=_make_mock_response("ok"))
 
-        with patch("lib.openai_shared.AsyncOpenAI", return_value=mock_client):
+        with captured_openai_clients(mock_client):
             from lib.text_backends.agnes import AgnesTextBackend
 
             backend = AgnesTextBackend(api_key="sk")
@@ -144,7 +129,7 @@ class TestGenerate:
         mock_client.chat.completions.create = AsyncMock(return_value=_make_mock_response(schema_response))
 
         with (
-            patch("lib.openai_shared.AsyncOpenAI", return_value=mock_client),
+            captured_openai_clients(mock_client),
             patch("instructor.from_openai") as from_openai,
         ):
             from lib.text_backends.agnes import AgnesTextBackend
@@ -176,7 +161,7 @@ class TestGenerate:
         )
 
         with (
-            patch("lib.openai_shared.AsyncOpenAI", return_value=mock_client),
+            captured_openai_clients(mock_client),
             patch("instructor.from_openai", return_value=mock_patched),
         ):
             from lib.text_backends.agnes import AgnesTextBackend
