+++ tests/unit/lib/video_backends/test_minimax_integration.py
@@ -9,6 +9,7 @@ import pytest
 from lib.pricing.lookup import lookup_pricing
 from lib.pricing.strategies import PricingParams, calculate_pricing
 from lib.providers import PROVIDER_MINIMAX, PROVIDER_OPENAI
+from tests.fakes import captured_openai_clients
 
 
 def _text_response(content: str = "ok", in_tok: int = 10, out_tok: int = 5) -> MagicMock:
@@ -58,7 +59,7 @@ class TestTextProviderBilling:
     """文本复用 OpenAI 后端必须以 'minimax' 记账，否则计费命中 USD。"""
 
     def test_provider_name_override(self):
-        with patch("lib.openai_shared.AsyncOpenAI"):
+        with captured_openai_clients():
             from lib.text_backends.openai import OpenAITextBackend
 
             b = OpenAITextBackend(api_key="k", model="MiniMax-M2.7", provider_name=PROVIDER_MINIMAX)
@@ -67,7 +68,7 @@ class TestTextProviderBilling:
     async def test_result_provider_is_minimax(self):
         mock_client = AsyncMock()
         mock_client.chat.completions.create = AsyncMock(return_value=_text_response("hi"))
-        with patch("lib.openai_shared.AsyncOpenAI", return_value=mock_client):
+        with captured_openai_clients(mock_client):
             from lib.text_backends.base import TextGenerationRequest
             from lib.text_backends.openai import OpenAITextBackend
